+++ sdk/src/main/AndroidManifest.xml
@@ -12,14 +12,14 @@
             android:theme="@style/Theme.ProcessOut.Default" />
 
         <activity
-            android:name=".ui.web.POCustomTabAuthorizationActivity"
+            android:name=".ui.web.customtab.POCustomTabAuthorizationActivity"
             android:configChanges="screenSize|smallestScreenSize|screenLayout|orientation|keyboard|keyboardHidden"
             android:exported="false"
             android:launchMode="singleTask"
             android:theme="@style/Theme.ProcessOut.Base.Transparent" />
 
         <activity
-            android:name=".ui.web.POCustomTabRedirectActivity"
+            android:name=".ui.web.customtab.POCustomTabRedirectActivity"
             android:exported="true"
             android:theme="@style/Theme.ProcessOut.Base.Transparent">
             <intent-filter
@@ -38,7 +38,7 @@
         </activity>
 
         <activity
-            android:name=".ui.web.POWebViewAuthorizationActivity"
+            android:name=".ui.web.webview.POWebViewAuthorizationActivity"
             android:configChanges="screenSize|smallestScreenSize|screenLayout|orientation|keyboard|keyboardHidden"
             android:exported="false"
             android:launchMode="singleTop"
