+++ sdk/src/main/kotlin/com/processout/sdk/ui/apm/POAlternativePaymentMethodCustomTabLauncher.kt
@@ -11,11 +11,11 @@ import com.processout.sdk.api.model.response.POAlternativePaymentMethodResponse
 import com.processout.sdk.api.network.ApiConstants
 import com.processout.sdk.core.ProcessOutActivityResult
 import com.processout.sdk.core.ProcessOutResult
-import com.processout.sdk.ui.web.CustomTabAuthorizationActivityContract
-import com.processout.sdk.ui.web.CustomTabConfiguration
+import com.processout.sdk.ui.web.customtab.CustomTabAuthorizationActivityContract
+import com.processout.sdk.ui.web.customtab.CustomTabConfiguration
 import com.processout.sdk.ui.web.WebAuthorizationDelegate
-import com.processout.sdk.ui.web.WebViewAuthorizationActivityLauncher
-import com.processout.sdk.ui.web.WebViewConfiguration
+import com.processout.sdk.ui.web.webview.WebViewAuthorizationActivityLauncher
+import com.processout.sdk.ui.web.webview.WebViewConfiguration
 
 class POAlternativePaymentMethodCustomTabLauncher private constructor() {
