+++ sdk/src/main/kotlin/com/processout/sdk/ui/threeds/PO3DSRedirectWebViewBuilder.kt
@@ -7,9 +7,9 @@ import com.processout.sdk.api.ProcessOut
 import com.processout.sdk.api.model.threeds.PO3DSRedirect
 import com.processout.sdk.api.network.ApiConstants
 import com.processout.sdk.core.ProcessOutResult
-import com.processout.sdk.ui.web.ProcessOutWebView
+import com.processout.sdk.ui.web.webview.ProcessOutWebView
 import com.processout.sdk.ui.web.WebAuthorizationDelegate
-import com.processout.sdk.ui.web.WebViewConfiguration
+import com.processout.sdk.ui.web.webview.WebViewConfiguration
 
 class PO3DSRedirectWebViewBuilder(
     private val activity: Activity
