+++ sdk/src/main/kotlin/com/processout/sdk/ui/web/webview/WebViewAuthorizationActivityLauncher.kt
@@ -1,4 +1,4 @@
-package com.processout.sdk.ui.web
+package com.processout.sdk.ui.web.webview
 
 import android.net.Uri
 import androidx.activity.ComponentActivity
@@ -7,6 +7,7 @@ import androidx.activity.result.ActivityResultLauncher
 import androidx.fragment.app.Fragment
 import com.processout.sdk.core.ProcessOutActivityResult
 import com.processout.sdk.core.ProcessOutResult
+import com.processout.sdk.ui.web.WebAuthorizationDelegate
 
 internal class WebViewAuthorizationActivityLauncher private constructor() {
