+++ sdk/src/main/kotlin/com/processout/sdk/ui/web/webview/WebViewConfiguration.kt
@@ -1,4 +1,4 @@
-package com.processout.sdk.ui.web
+package com.processout.sdk.ui.web.webview
 
 import android.net.Uri
 import android.os.Parcelable
